Fix substitution bug.
There may be a fancier way to implement this, where the library still
makes the recursive substitution calls, but this approach seems both
flexible and light weight: in the cases where variables are
encountered, or where post processing is needed, the user defines
'substHookM' to return a hook function that takes over the
substitution at that point.  In our code, the "variable" case and the
"post processing" case are combined for 'VNeut'.

Here's the instance:

  instance SubstM SpireM Value Value where
    substHookM (VNeut nm fs) = Just f
      where
      f nm' e = do
        fs' <- substM nm' e fs
        let head = if nm == nm' then e else vVar nm
        elims head fs'
    substHookM _ = Nothing

One possible downside is that the parallel substitution 'substsM'
seems hard to define now. But we weren't using that anyway.