Fix deBruijn capture bug for 'Bind' binders.
But similar changes probably need be made for other types with smart
constructors that use 'close' (and for other 'GenBind' based types in
particular).  A simpler solution is probably to just fix the Unbound
library so it correctly handles free deBruijn vars in 'close' and
'open' (by weakening and strengthening).